correct error

+++ lib/common.sh
@@ -111,10 +111,11 @@ removeWorkspace() {
     execute rm -rf "${workspace}"
     files_left=$(execute -n ls -R1|wc -l)
     debug "files left after executing \"rm -rf ${workspace}\": ${files_left}"
-    debug "changing access rights for left files"
-    execute chmod -R u+w "${workspace}"
-    execute rm -rf "${workspace}"
-
+    if [[ ${files_left} -ge 1 ]] ; then
+        debug "changing access rights for left files"
+        execute chmod -R u+w "${workspace}"
+        execute rm -rf "${workspace}"
+    fi
     return
 }
